real-estate-analyzer/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.2
httpx==0.28.1
beautifulsoup4==4.13.4
sqlalchemy==2.0.41
aiosqlite==0.21.0
pydantic==2.11.3
pydantic-settings==2.9.1
pandas==2.2.3
numpy==2.2.6
diskcache==5.6.3
python-dotenv==1.1.0
lxml==5.4.0
reportlab==4.4.0
anthropic==0.52.0
supabase>=2.5.0
python-jose[cryptography]>=3.3.0
